spasm_LU --> spasm_GPLU

+++ src/spasm_util.c
@@ -1,5 +1,7 @@
 #include <sys/time.h>
 #include <err.h>
+#include <inttypes.h>
+
 #include "spasm.h"
 
 int spasm_get_num_threads() {
@@ -33,7 +35,7 @@ int spasm_nnz(const spasm * A) {
 /* return a string representing n in 4 bytes */
 void spasm_human_format(int64_t n, char *target) {
 	if (n < 1000) {
-		sprintf(target, "%lld", n);
+		sprintf(target, "%d", (int) n);
 		return;
 	}
 	if (n < 1000000) {
